backend/requirements.txt
Django==1.11.6
djangorestframework==3.7.1
pytz==2017.2
coreapi==2.3.3
psycopg2==2.7.3
networkx==2.0